+++ esp-config/src/bin/esp-config/main.rs
@@ -8,6 +8,7 @@ use std::{
 use clap::Parser;
 use env_logger::{Builder, Env};
 use esp_config::{ConfigOption, Value};
+use esp_metadata_generated::Chip;
 use serde::Deserialize;
 use toml_edit::{DocumentMut, Formatted, Item, Table};
 
@@ -21,12 +22,16 @@ struct Args {
     path: Option<PathBuf>,
 
     /// Chip
-    #[arg(short = 'C', long)]
-    chip: Option<esp_metadata::Chip>,
+    #[arg(short = 'C', long, value_parser = chip_from_str)]
+    chip: Option<Chip>,
 
     /// Config file - using `config.toml` by default
     #[arg(short = 'c', long)]
     config_file: Option<String>,
+}
+
+fn chip_from_str(str: &str) -> Result<Option<Chip>, String> {
+    Chip::from_str(str).map(Some)
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -178,7 +183,7 @@ fn apply_config(
 
 fn parse_configs(
     path: &Path,
-    chip_from_args: Option<esp_metadata::Chip>,
+    chip_from_args: Option<Chip>,
     config_toml_path: &PathBuf,
 ) -> Result<(bool, Vec<CrateConfig>), Box<dyn Error>> {
     let mut hint_about_configs = false;
@@ -244,37 +249,30 @@ fn parse_configs(
         let mut chip = None;
         for pkg in &meta.root_package().unwrap().dependencies {
             if pkg.name == "esp-hal" {
-                let possible_chip_feature_matches: Vec<esp_metadata::Chip> = pkg
-                    .features
-                    .iter()
-                    .flat_map(|f| esp_metadata::Chip::from_str(f))
-                    .collect::<Vec<esp_metadata::Chip>>();
-                chip = possible_chip_feature_matches.first().cloned();
+                chip = pkg.features.iter().flat_map(|f| Chip::from_str(f)).next();
             }
         }
         chip
     };
 
-    // the "ESP_CONFIG_CHIP" hint env-var if present
+    // the "ESP_CONFIG_CHIP" hint env-var if present (could be set in a config.toml)
     let chip_from_config = || {
         envs.get("ESP_CONFIG_CHIP")
-            .and_then(|chip_str| clap::ValueEnum::from_str(chip_str, true).ok())
+            .and_then(|chip_str| Chip::from_str(&chip_str.to_lowercase()).ok())
     };
 
     // - if given as a parameter, use it
     // - if there is a hint in the config.toml, use it
     // - if we can infer it from metadata, use it
     // otherwise, fail
-    let chip = chip_from_args
+    let Some(chip) = chip_from_args
         .or_else(chip_from_config)
-        .or_else(chip_from_meta);
-
-    if chip.is_none() {
+        .or_else(chip_from_meta)
+    else {
         return Err("No chip given or inferred. Try using the `--chip` argument.".into());
-    }
+    };
 
     let mut configs = Vec::new();
-    let chip = esp_metadata_generated::Chip::from_str(chip.unwrap().as_ref()).unwrap();
     let features = vec![];
     for krate in meta.packages {
         let maybe_cfg = krate.manifest_path.parent().unwrap().join("esp_config.yml");