+++ esp-radio/src/wifi/mod.rs
@@ -19,8 +19,6 @@ use num_derive::FromPrimitive;
 pub(crate) use os_adapter::*;
 use portable_atomic::{AtomicUsize, Ordering};
 use procmacros::BuilderLite;
-#[cfg(feature = "serde")]
-use serde::{Deserialize, Serialize};
 #[cfg(all(feature = "smoltcp", feature = "unstable"))]
 #[cfg_attr(docsrs, doc(cfg(feature = "unstable")))]
 use smoltcp::phy::{Device, DeviceCapabilities, RxToken, TxToken};
@@ -151,10 +149,17 @@ use crate::sys::{
     },
 };
 
+const _: () = {
+    // make sure we know all the auth modes the driver knows
+    core::assert!(
+        include::wifi_auth_mode_t_WIFI_AUTH_MAX == 17,
+        "Make sure all auth-methods known by the driver are known by us."
+    )
+};
+
 /// Supported Wi-Fi authentication methods.
 #[derive(Copy, Clone, Debug, Default, Eq, PartialEq, PartialOrd)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 #[non_exhaustive]
 pub enum AuthMethod {
     /// No authentication (open network).
@@ -184,12 +189,39 @@ pub enum AuthMethod {
 
     /// WLAN Authentication and Privacy Infrastructure (WAPI).
     WapiPersonal,
+
+    /// Opportunistic Wireless Encryption (OWE)
+    Owe,
+
+    /// WPA3 Enterprise Suite B 192-bit Encryption
+    Wpa3EntSuiteB192Bit,
+
+    /// This authentication mode will yield same result as [AuthMethod::Wpa3Personal] and is not
+    /// recommended to be used. It will be deprecated in future, please use
+    /// [AuthMethod::Wpa3Personal] instead.
+    Wpa3ExtPsk,
+
+    /// This authentication mode will yield same result as [AuthMethod::Wpa3Personal] and is not
+    /// recommended to be used. It will be deprecated in future, please use
+    /// [AuthMethod::Wpa3Personal] instead.
+    Wpa3ExtPskMixed,
+
+    /// WiFi DPP / Wi-Fi Easy Connect
+    Dpp,
+
+    /// WPA3-Enterprise Only Mode
+    Wpa3Enterprise,
+
+    /// WPA3-Enterprise Transition Mode
+    Wpa2Wpa3Enterprise,
+
+    /// WPA-Enterprise security
+    WpaEnterprise,
 }
 
 /// Supported Wi-Fi protocols.
 #[derive(Debug, Default, PartialOrd, EnumSetType)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 #[non_exhaustive]
 pub enum Protocol {
     /// 802.11b protocol.
@@ -232,7 +264,6 @@ impl Protocol {
 /// Secondary Wi-Fi channels.
 #[derive(Clone, Debug, Default, Eq, PartialEq, PartialOrd)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 pub enum SecondaryChannel {
     // TODO: Need to extend that for 5GHz
     /// No secondary channel (default).
@@ -248,7 +279,6 @@ pub enum SecondaryChannel {
 
 /// Access point country information.
 #[derive(Clone, Debug, Default, PartialEq, Eq)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 pub struct Country([u8; 2]);
 
 impl Country {
@@ -292,7 +322,6 @@ impl defmt::Format for Country {
 /// Information about a detected Wi-Fi access point.
 #[derive(Clone, Debug, Default, Eq, PartialEq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 #[non_exhaustive]
 pub struct AccessPointInfo {
     /// The SSID of the access point.
@@ -321,7 +350,6 @@ pub struct AccessPointInfo {
 
 /// Configuration for a Wi-Fi access point.
 #[derive(BuilderLite, Clone, Eq, PartialEq)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 pub struct AccessPointConfig {
     /// The SSID of the access point.
     #[builder_lite(reference)]
@@ -453,7 +481,6 @@ pub enum ScanMethod {
 
 /// Client configuration for a Wi-Fi connection.
 #[derive(BuilderLite, Clone, Eq, PartialEq)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 pub struct ClientConfig {
     /// The SSID of the Wi-Fi network.
     #[builder_lite(reference)]
@@ -588,7 +615,6 @@ impl defmt::Format for ClientConfig {
 /// Configuration for EAP-FAST authentication protocol.
 #[derive(Clone, Debug, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 #[cfg(feature = "wifi-eap")]
 #[instability::unstable]
 pub struct EapFastConfig {
@@ -603,7 +629,6 @@ pub struct EapFastConfig {
 /// Phase 2 authentication methods
 #[derive(Debug, Clone, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 #[cfg(feature = "wifi-eap")]
 #[instability::unstable]
 pub enum TtlsPhase2Method {
@@ -652,7 +677,6 @@ type CertificateAndKey = (&'static [u8], &'static [u8], Option<&'static [u8]>);
 
 /// Configuration for an EAP (Extensible Authentication Protocol) client.
 #[derive(BuilderLite, Clone, PartialEq, Eq)]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 #[cfg(feature = "wifi-eap")]
 #[instability::unstable]
 pub struct EapClientConfig {
@@ -875,7 +899,6 @@ impl Default for EapClientConfig {
 /// Introduces Wi-Fi configuration options.
 #[derive(EnumSetType, Debug, PartialOrd)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 #[non_exhaustive]
 pub enum Capability {
     /// The device operates as a client, connecting to an existing network.
@@ -894,7 +917,6 @@ pub enum Capability {
 #[allow(clippy::large_enum_variant)]
 #[derive(Clone, Debug, PartialEq, Eq, Default)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 #[non_exhaustive]
 pub enum ModeConfig {
     /// No configuration (default).
@@ -912,7 +934,6 @@ pub enum ModeConfig {
 
     /// EAP client configuration for enterprise Wi-Fi.
     #[cfg(feature = "wifi-eap")]
-    #[cfg_attr(feature = "serde", serde(skip))]
     EapClient(EapClientConfig),
 }
 
@@ -951,6 +972,18 @@ impl AuthMethodExt for AuthMethod {
             AuthMethod::Wpa3Personal => include::wifi_auth_mode_t_WIFI_AUTH_WPA3_PSK,
             AuthMethod::Wpa2Wpa3Personal => include::wifi_auth_mode_t_WIFI_AUTH_WPA2_WPA3_PSK,
             AuthMethod::WapiPersonal => include::wifi_auth_mode_t_WIFI_AUTH_WAPI_PSK,
+            AuthMethod::Owe => include::wifi_auth_mode_t_WIFI_AUTH_OWE,
+            AuthMethod::Wpa3EntSuiteB192Bit => include::wifi_auth_mode_t_WIFI_AUTH_WPA3_ENT_192,
+            AuthMethod::Wpa3ExtPsk => include::wifi_auth_mode_t_WIFI_AUTH_WPA3_EXT_PSK,
+            AuthMethod::Wpa3ExtPskMixed => {
+                include::wifi_auth_mode_t_WIFI_AUTH_WPA3_EXT_PSK_MIXED_MODE
+            }
+            AuthMethod::Dpp => include::wifi_auth_mode_t_WIFI_AUTH_DPP,
+            AuthMethod::Wpa3Enterprise => include::wifi_auth_mode_t_WIFI_AUTH_WPA3_ENTERPRISE,
+            AuthMethod::Wpa2Wpa3Enterprise => {
+                include::wifi_auth_mode_t_WIFI_AUTH_WPA2_WPA3_ENTERPRISE
+            }
+            AuthMethod::WpaEnterprise => include::wifi_auth_mode_t_WIFI_AUTH_WPA_ENTERPRISE,
         }
     }
 
@@ -965,7 +998,25 @@ impl AuthMethodExt for AuthMethod {
             include::wifi_auth_mode_t_WIFI_AUTH_WPA3_PSK => AuthMethod::Wpa3Personal,
             include::wifi_auth_mode_t_WIFI_AUTH_WPA2_WPA3_PSK => AuthMethod::Wpa2Wpa3Personal,
             include::wifi_auth_mode_t_WIFI_AUTH_WAPI_PSK => AuthMethod::WapiPersonal,
-            _ => unreachable!(),
+            include::wifi_auth_mode_t_WIFI_AUTH_OWE => AuthMethod::Owe,
+            include::wifi_auth_mode_t_WIFI_AUTH_WPA3_ENT_192 => AuthMethod::Wpa3EntSuiteB192Bit,
+            include::wifi_auth_mode_t_WIFI_AUTH_WPA3_EXT_PSK => AuthMethod::Wpa3ExtPsk,
+            include::wifi_auth_mode_t_WIFI_AUTH_WPA3_EXT_PSK_MIXED_MODE => {
+                AuthMethod::Wpa3ExtPskMixed
+            }
+            include::wifi_auth_mode_t_WIFI_AUTH_DPP => AuthMethod::Dpp,
+            include::wifi_auth_mode_t_WIFI_AUTH_WPA3_ENTERPRISE => AuthMethod::Wpa3Enterprise,
+            include::wifi_auth_mode_t_WIFI_AUTH_WPA2_WPA3_ENTERPRISE => {
+                AuthMethod::Wpa2Wpa3Enterprise
+            }
+            include::wifi_auth_mode_t_WIFI_AUTH_WPA_ENTERPRISE => AuthMethod::WpaEnterprise,
+            // we const-assert we know all the auth-methods the wifi driver knows and it shouldn't
+            // return anything else.
+            //
+            // In fact from observation the drivers will return
+            // `wifi_auth_mode_t_WIFI_AUTH_OPEN` if the method is unsupported (e.g. any WPA3 in our
+            // case, since the supplicant isn't compiled to support it)
+            _ => AuthMethod::None,
         }
     }
 }
@@ -973,7 +1024,6 @@ impl AuthMethodExt for AuthMethod {
 /// Wi-Fi Mode (Sta and/or Ap)
 #[derive(Debug, Clone, Copy, PartialEq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-#[cfg_attr(feature = "serde", derive(Deserialize, Serialize))]
 #[non_exhaustive]
 pub enum WifiMode {
     /// Station mode.