+++ qa-test/src/bin/gpio_interrupt_latency.rs
@@ -21,11 +21,10 @@
 use embassy_executor::Spawner;
 use embassy_sync::{blocking_mutex::raw::CriticalSectionRawMutex, signal::Signal};
 use esp_backtrace as _;
-#[cfg(target_arch = "riscv32")]
-use esp_hal::interrupt::software::SoftwareInterruptControl;
 use esp_hal::{
     clock::CpuClock,
     gpio::{Flex, OutputConfig},
+    interrupt::software::SoftwareInterruptControl,
     timer::timg::TimerGroup,
 };
 
@@ -49,14 +48,10 @@ async fn main(spawner: Spawner) {
     enc_b.apply_output_config(&OutputConfig::default());
     enc_b.set_output_enable(true);
     enc_b.set_input_enable(true);
-    #[cfg(target_arch = "riscv32")]
+
     let sw_int = SoftwareInterruptControl::new(p.SW_INTERRUPT);
     let timg0 = TimerGroup::new(p.TIMG0);
-    esp_rtos::start(
-        timg0.timer0,
-        #[cfg(target_arch = "riscv32")]
-        sw_int.software_interrupt0,
-    );
+    esp_rtos::start(timg0.timer0, sw_int.software_interrupt0);
 
     spawner.must_spawn(toggle(enc_a_clone, enc_b_clone, &SIGNAL));
     spawner.must_spawn(wait(enc_a, enc_b, &SIGNAL));